config: Move Config struct to new state.h
